+++ pystac_client/item_search.py
@@ -745,11 +745,13 @@ class ItemSearch:
     # not caching these, since they're cached in the implementation
 
     def get_item_collections(self) -> Iterator[ItemCollection]:
-        """DEPRECATED. Use :meth:`ItemSearch.item_collections` instead.
+        """DEPRECATED
+
+        .. deprecated:: 0.4.0
+            Use :meth:`ItemSearch.pages` instead.
 
         Yields:
-            ItemCollection : a group of Items matching the search criteria within an
-            ItemCollection
+            ItemCollection : a group of Items matching the search criteria.
         """
         warnings.warn(
             "get_item_collections() is deprecated, use pages() instead",
@@ -758,21 +760,26 @@ class ItemSearch:
         return self.pages()
 
     def item_collections(self) -> Iterator[ItemCollection]:
-        """Iterator that yields ItemCollection objects.  Each ItemCollection is
-        a page of results from the search.
+        """DEPRECATED
+
+        .. deprecated:: 0.5.0
+            Use :meth:`ItemSearch.pages` instead.
 
         Yields:
             ItemCollection : a group of Items matching the search criteria within an
             ItemCollection
         """
         warnings.warn(
-            "'item_collections()' is deprecated, use 'pages()' instead",
+            "item_collections() is deprecated, use pages() instead",
             DeprecationWarning,
         )
         return self.pages()
 
     def get_items(self) -> Iterator[Item]:
-        """DEPRECATED. Use :meth:`ItemSearch.items` instead.
+        """DEPRECATED.
+
+        .. deprecated:: 0.4.0
+            Use :meth:`ItemSearch.items` instead.
 
         Yields:
             Item : each Item matching the search criteria
@@ -784,8 +791,7 @@ class ItemSearch:
         return self.items()
 
     def get_all_items(self) -> ItemCollection:
-        """
-        Get the matching items as a :ref:`pystac.ItemCollection`.
+        """DEPRECATED
 
         .. deprecated:: 0.4.0
            Use :meth:`ItemSearch.item_collection` instead.
@@ -794,23 +800,22 @@ class ItemSearch:
             item_collection : ItemCollection
         """
         warnings.warn(
-            "get_all_items is deprecated, use 'item_collection' instead.",
+            "get_all_items() is deprecated, use item_collection() instead.",
             DeprecationWarning,
         )
         return self.item_collection()
 
     def get_all_items_as_dict(self) -> Dict[str, Any]:
-        """Get items as a FeatureCollection dictionary.
+        """DEPRECATED
 
-        Convenience method that gets all items from all pages, up to
-        the number provided by the max_items parameter, and returns an array of
-        dictionaries.
+        .. deprecated:: 0.4.0
+           Use :meth:`ItemSearch.item_collection_as_dict` instead.
 
         Return:
             Dict : A GeoJSON FeatureCollection
         """
         warnings.warn(
-            "'get_all_items_as_dict' is deprecated, use 'item_collection_as_dict' "
+            "get_all_items_as_dict() is deprecated, use item_collection_as_dict() "
             "instead.",
             DeprecationWarning,
         )